+++ src/integrationTest/kotlin/me/elgregos/reakteves/infrastructure/event/ReactorEventBusTest.kt
@@ -4,7 +4,7 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import me.elgregos.reakteves.config.BaseIntegrationTest
 import me.elgregos.reakteves.domain.event.Event
-import me.elgregos.reakteves.domain.fakeCreatedEvent
+import me.elgregos.reakteves.domain.event.fakeCreatedEvent
 import me.elgregos.reakteves.libs.genericObjectMapper
 import org.awaitility.kotlin.await
 import org.springframework.beans.factory.annotation.Autowired
