+++ src/test/kotlin/me/elgregos/reakteves/domain/event/AggregateTest.kt
@@ -4,8 +4,6 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import io.mockk.every
 import io.mockk.mockk
-import me.elgregos.reakteves.domain.FakeEvent
-import me.elgregos.reakteves.domain.fakeCreatedEvent
 import me.elgregos.reakteves.libs.genericObjectMapper
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
