+++ src/test/kotlin/me/elgregos/reakteves/infrastructure/event/EventEntityTest.kt
@@ -2,9 +2,8 @@ package me.elgregos.reakteves.infrastructure.event
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
-import me.elgregos.reakteves.domain.FakeEvent
-import me.elgregos.reakteves.domain.fakeCreatedEvent
-import me.elgregos.reakteves.infrastructure.FakeEventEntity
+import me.elgregos.reakteves.domain.event.FakeEvent
+import me.elgregos.reakteves.domain.event.fakeCreatedEvent
 import me.elgregos.reakteves.libs.genericObjectMapper
 import java.util.*
 import kotlin.test.Test
