+++ src/testFixtures/kotlin/me/elgregos/reakteves/infrastructure/event/FakeEventEntity.kt
@@ -1,8 +1,7 @@
-package me.elgregos.reakteves.infrastructure
+package me.elgregos.reakteves.infrastructure.event
 
 import com.fasterxml.jackson.databind.JsonNode
-import me.elgregos.reakteves.domain.FakeEvent
-import me.elgregos.reakteves.infrastructure.event.EventEntity
+import me.elgregos.reakteves.domain.event.FakeEvent
 import me.elgregos.reakteves.libs.nowUTC
 import org.springframework.data.relational.core.mapping.Table
 import java.time.LocalDateTime
