+++ src/testFixtures/kotlin/me/elgregos/reakteves/infrastructure/event/FakeEventRepository.kt
@@ -1,0 +1,6 @@
+package me.elgregos.reakteves.infrastructure.event
+
+import me.elgregos.reakteves.domain.event.FakeEvent
+import java.util.*
+
+interface FakeEventRepository: EventEntityRepository<FakeEventEntity, FakeEvent, UUID>
